some ui refactoring
