+++ app/src/main/java/com/matsyuk/mobiusclean/clean/ui/main/views/MainActivity.java
@@ -7,7 +7,7 @@ import android.widget.Button;
 
 import com.matsyuk.mobiusclean.R;
 import com.matsyuk.mobiusclean.clean.dagger.ComponentManager;
-import com.matsyuk.mobiusclean.clean.ui.app_lock.AppLockActivity;
+import com.matsyuk.mobiusclean.clean.ui.app_lock.views.AppLockActivity;
 import com.matsyuk.mobiusclean.clean.ui.main.presenters.IMainPresenter;
 import com.matsyuk.mobiusclean.clean.ui.start.views.StartActivity;
