+++ app/src/main/res/layout/ac_app_lock.xml
@@ -7,9 +7,9 @@
     android:paddingLeft="16dp"
     android:paddingRight="16dp">
 
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="App lock"/>
+    <FrameLayout
+        android:id="@+id/start_container"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"/>
 
 </LinearLayout>
